+++ MoMoney.Core/ViewModels/EditTransactionViewModel.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.ObjectModel;
-using CommunityToolkit.Mvvm.Input;
+﻿using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.ComponentModel;
 using MvvmHelpers;
 using MoMoney.Core.Models;